feat: 탈퇴한 회원 익명처리 기능추가 (#58)
