+++ api-module/src/main/java/org/devridge/api/domain/community/mapper/CommunityMapper.java
@@ -3,6 +3,7 @@ package org.devridge.api.domain.community.mapper;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
+import lombok.RequiredArgsConstructor;
 import org.devridge.api.domain.community.dto.request.CreateCommunityRequest;
 import org.devridge.api.domain.community.dto.response.CommunityDetailResponse;
 import org.devridge.api.domain.community.dto.response.CommunityListResponse;
@@ -12,13 +13,17 @@ import org.devridge.api.domain.member.entity.Member;
 import org.springframework.stereotype.Component;
 
 @Component
+@RequiredArgsConstructor
 public class CommunityMapper {
+
+    private final MemberInfoMapper memberInfoMapper;
 
     public CommunityDetailResponse toCommunityDetailResponse(Community community) {
         Member member = community.getMember();
+        MemberInfoResponse memberInfoResponse = memberInfoMapper.toMemberInfoResponse(member);
         return CommunityDetailResponse.builder()
                 .communityId(community.getId())
-                .memberInfoResponse(toMemberInfoResponse(member))
+                .memberInfoResponse(memberInfoResponse)
                 .title(community.getTitle())
                 .content(community.getContent())
                 .likeCount(community.getLikeCount())
@@ -28,15 +33,6 @@ public class CommunityMapper {
                 .updatedAt(community.getUpdatedAt())
                 .hashtags(toHashtags(community))
                 .build();
-    }
-
-    public MemberInfoResponse toMemberInfoResponse(Member member) {
-        return MemberInfoResponse.builder()
-            .memberId(member.getId())
-            .nickName(member.getNickname())
-            .profileImageUrl(member.getProfileImageUrl())
-            .introduction(member.getIntroduction())
-            .build();
     }
 
     public List<String> toHashtags(Community community) {
